Crate gphoto

Structs

Abilities
Describes the abilities of a device.

Camera
A structure representing a camera connected to the system.

CameraFile
A file stored on a camera's storage.

Context
A libgphoto2 library context.

Error
An error type for working with gphoto2.

FileMedia
Media stored as a local file.

LibraryVersion
A structure that describes the version of the libgphoto2 library.

Port
A structure describing a port.

Storage
Structure containing information about a camera's storage.

Enums

AccessType
Types of access permissions.

CameraOperation
Operations that can be performed on a device.

DeviceType
Types of devices.

DriverStatus
Stability of camera driver.

ErrorKind
Types of errors reported by gphoto2.

FileOperation
Operations that can be performed on files on a device's storage.

FilesystemType
Types of filesystem hierarchies.

FolderOperation
Operations that can be performed on folders on a device's storage.

PortType
Types of ports.

StorageType
Types of storage hardware.

Traits

Media
A trait for types that can store media.

Functions

libgphoto2_version
Returns a structure with the version of the libgphoto2 C library.

Type Definitions

Result
A specialized Result type for working with gphoto2.