Struct gphoto::CameraFile

pub struct CameraFile {
    // some fields omitted
}

A file stored on a camera's storage.

Methods

impl CameraFile

fn directory(&self) -> Cow<str>

Returns the directory that the file is stored in.

fn basename(&self) -> Cow<str>

Returns the name of the file without the directory.