Struct gphoto::LibraryVersion

pub struct LibraryVersion {
    // some fields omitted
}

A structure that describes the version of the libgphoto2 library.

The structure describes not only the version number, but also the state of several compile-time configuration options of the libgphoto2 library.

Example

let version = gphoto::libgphoto2_version();
println!("libgphoto2 {} {} {} {} {}",
         version.version(),
         version.camlibs(),
         version.compiler(),
         version.ltdl(),
         version.exif());

This may output something like the following:

libgphoto2 2.5.7 all camlibs clang ltdl no EXIF

Methods

impl LibraryVersion

fn version(&self) -> &str

Returns the libgphoto2 version number.

The version number is a string representing the version number, e.g., "2.4.16".

fn camlibs(&self) -> &str

Describes which camlibs were compiled with libgphoto2.

fn compiler(&self) -> &str

Returns the name of the C compiler that was used to compile libgphoto2.

fn ltdl(&self) -> &str

Returns a string describing whether libgphoto2 was compiled for portable loading of camlibs.

fn exif(&self) -> &str

Returns a string describing whether libgphoto2 was compiled with support for handling EXIF metadata.